Working to fix "NoOutsideLoads" b.c. issues
What I thought before were problems with the solution appear to be issues
with Dirichlet b.c.'s that are too close to the flexing portions of the plate.

There is no solution convergence when I increase padding in the case of
"NoOutsideLoads".

I think that this is because of how I enforce periodic boundary conditions,
and possibly because of discontinuities in Te on the boundaries.

These edits are meant to try to approach the b.c. by removing the periodic
part and testing how the grid responds.