Remove "NoOutsideLoads" boundary condition option
This boundary condition's implementation really was a way to avoid including
the full elastic thickness array of the region of importance. It introduced
boundary-related issues in the solutions. I have therefore removed it, and
have plans to allow the elastic thickness and load arrays to be different
sizes, with the load arrays' location within the elastic thickness array
defined by the user.