Fixed Periodic B.C.: IMPORTANT for long-term issues; read notes.
The periodic boundary conditions were also not implemented with "np.roll",
and had the values given in the wrong order, to boot. Either of these issues
affects only the variable Te case, of course: no problem for constnat Te
because every item in that array is the same. This is why I had issues with
mirror boundary conditions finding a discontinuity in the 2D code, when I
generated the ad hoc boundaries by just padding and extending the array. This
is likely why these boundary conditions behaved as poorly as Dirichlet ones:
they were forced to go to 0! And therefore I started seeing "ringing"
(oscillation) around the boundaries.